requirements.txt
boto3>=1.26.0
click>=8.1.3
pyyaml>=6.0
python-dotenv>=1.0.0
